% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTrajector.R
\name{clusterTraj}
\alias{clusterTraj}
\title{Cluster the nonlinear trajector}
\usage{
clusterTraj(
  dataset,
  phemeta,
  metavar,
  scale = "log",
  span = 0.7,
  na.rm = T,
  cut,
  ...
)
}
\arguments{
\item{dataset}{dataframe,microbe or lipid datset, row:sample id ,col: feature}

\item{phemeta}{dataframe,phenotye dataset}

\item{metavar}{character,the intersting feature}

\item{scale}{optinal character, log or int}

\item{span}{numerical value}

\item{na.rm}{logical value indicating whether NA values should be stripped before the computation proceeds.}

\item{cut}{numerical value}

\item{...}{urther arguments passed to or from other methods}
}
\value{
list
}
\description{

}
\examples{


}
